requirements.txt
backports.ssl-match-hostname==3.4.0.2
base58==0.2.1
bottle==0.12.8
click==3.3
gevent==1.0.1
greenlet==0.4.5
libnacl==1.4.0
lmdb==0.84
msgpack-python==0.4.6
pycrypto==2.6.1
PyDispatcher==2.0.5
PyYAML==3.11
pyscrypt==1.6.2
requests==2.4.3
six==1.9.0
ujson==1.33
ws4py==0.3.4
wsaccel==0.6.2

